[lint package refactor] Create a file for the utility functions
